New Enumeration class called Note
To make it easy to compose chords without needing to store all the souns in the app (reducing the app size). A chord can be composed by 3 or 4 notes.

ReadSongActivity: Replaced the MediaPlayer by SoundPool to handle 4 notes at a time.

TODO: In the ChordAdapter, show all available chords, pending to check the sounds.